Moved VisualCodeGraph to visualization.py
