+++ visualization.py
@@ -1,5 +1,5 @@
 import plotly.express as px
-from napkin_graph import CodeBase, CodeGraph, VisualCodeGraph
+from napkin_graph import CodeBase, CodeGraph
 """
 The structure of graph_dict is as follows:
 dict[File, list[dict[Class, list[Function]]]]
@@ -16,6 +16,78 @@ There are no repeated names for files, classes, functions
     ],
 }
 """
+
+class VisualCodeGraph():
+    """ Class for generating visualizations of the code graph
+    """
+
+    def __init__(self, codegraph: CodeGraph) -> None:
+        self.codegraph = codegraph
+        self.graph_dict = {}
+        # LonelyFunctions are functions directly written into a class
+        self.lonelyFunctions = {}
+        self.graph_dict = self._create_graph_dict()
+
+    def get_graph_dict(self, print: bool = False):
+        if self.graph_dict:
+            return self.graph_dict
+        else:
+            self.graph_dict = self._create_graph_dict()
+        if print:
+            print(f"Graph Dictionary:\n\t{self.graph_dict}")
+        return self.graph_dict
+
+    def get_lonely_functions(self):
+        return self.lonelyFunctions
+
+    def _create_graph_dict(self) -> 'dict[File, list[dict[Class, list[Function]]]]':
+        """ Creates a dictionary representation of the graph
+            where keys are files and values are a list of dictionaries
+            representing the classes and their functions
+
+            This function also populates self.lonelyFunctions
+
+            Example Output:
+            {
+                'file1.py': [
+                                {
+                                    'Class1': ['Function1', 'Function2']
+                                },
+                                {
+                                    'Class2': ['Function3', 'Function4']
+                                }
+                ]
+            }
+        Returns:
+            dict[File, list[dict[Class, list[Function]]]]: The dictionary representation of the graph
+        """
+        graph_dict = {}
+        nodes = self.codegraph.nodes
+        for node in nodes:
+            if node.is_file():
+                file = node
+                graph_dict[file] = []
+                # Get the connected nodes
+                connected_nodes = self.codegraph.find_connected_nodes(file)
+                for connected_node in connected_nodes:
+                    if connected_node.is_class():
+                        class_dict = {}
+                        class_dict[connected_node] = []
+                        # Get the connected nodes
+                        connected_nodes2 = self.codegraph.find_connected_nodes(
+                            connected_node)
+                        for connected_node2 in connected_nodes2:
+                            if connected_node2 == None:
+                                continue
+                            elif connected_node2.is_function():
+                                class_dict[connected_node].append(
+                                    connected_node2)
+                        graph_dict[file].append(class_dict)
+                    elif connected_node.is_function():
+                        print(
+                            f"Found lonely Function {connected_node.name} -> File {file.name}")
+                        self.lonelyFunctions[connected_node] = file
+        return graph_dict
 
 
 class KnowledgeGraphVisualizer():
